requirements_v1.txt
# requirements_v1.txt
# Dependencies for Chat Logger v1.0 with Atomic Prompting
requests
python-dotenv
tiktoken